+++ socketsecurity/__init__.py
@@ -1,3 +1,3 @@
 __author__ = 'socket.dev'
-__version__ = '2.2.68'
+__version__ = '2.2.69'
 USER_AGENT = f'SocketPythonCLI/{__version__}'